add trait for costs remove method from controller, add trait on controller
