+++ src/fredi/AppBundle/Resources/views/App/dockbar.html.twig
@@ -4,7 +4,7 @@
             <ul class="nav navbar-nav">
                 <li><a href="{{ path('fredi_app_homepage') }}">Accueil</a></li>
                 <li><a href="{{ path('costline_new', {'associationUniqueId': associationUniqueId}) }}">Enregistrer une note de frais</a></li>
-                <li><a href="{{ path('cost_index') }}">Consulter le bordereau</a></li>
+                <li><a href="{{ path('costline_index', { 'associationUniqueId': associationUniqueId }) }}">Consulter le bordereau</a></li>
                 <li><a href="#">Générer le PDF</a></li>
                 <li class="dropdown">
                     <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="username">{{ app.user.username}}</span></a>
